requirements.txt
flask[async]
Flask-Login
python-dotenv
aiohttp
pymongo
flask_pymongo
waitress
Flask-WTF
requests
Flask-Limiter[mongodb]
pywebpush
py-vapid
cachetools
Flask-CORS
colorlog
